requirements.txt
fastapi
uvicorn
pyzmq
tornado